Most of the reports are done. Special-action-reports still to go.
STILL, DO NOT UPDATE ON THE ROOTSERVER YET!!!

+++ src/reports/OtherReport.cpp
@@ -13,12 +13,12 @@ void OtherReport::setShips(std::string ships) {
 	this->ships = ships;
 }
 
-void OtherReport::setRes(unsigned long res0,
-					   unsigned long res1,
-					   unsigned long res2,
-					   unsigned long res3,
-					   unsigned long res4,
-					   unsigned long res5) {
+void OtherReport::setRes(double res0,
+					   double res1,
+					   double res2,
+					   double res3,
+					   double res4,
+					   double res5) {
 	this->res0 = res0;
 	this->res1 = res1;
 	this->res2 = res2;
